restructure- move reducer into separate file
